+++ packages/base-ui/src/layouts/docs/page/index.tsx
@@ -1,9 +1,6 @@
-import type { ComponentProps, ReactNode } from 'react';
-import { cn } from '@/utils/cn';
-import { buttonVariants } from '@/components/ui/button';
-import { Edit, Text } from 'lucide-react';
-import { I18nLabel } from '@/contexts/i18n';
+'use client';
 import {
+<<<<<<< HEAD
   type BreadcrumbProps,
   type FooterProps,
   PageBreadcrumb,
@@ -17,131 +14,152 @@ import * as TocDefault from '@/components/toc/default';
 import * as TocClerk from '@/components/toc/clerk';
 import { TOCProvider, TOCScrollArea } from '@/components/toc';
 import { ChildrenRenderer, renderer, type Renderer } from '@/utils/renderer';
+=======
+  type ComponentProps,
+  createContext,
+  type FC,
+  type ReactNode,
+  use,
+  useEffect,
+  useState,
+} from 'react';
+import { cn } from '@/utils/cn';
+import { I18nLabel, useI18n } from '@/contexts/i18n';
+import {
+  TOC,
+  TOCPopover,
+  TOCProvider,
+  type TOCProviderProps,
+  type TOCPopoverProps,
+  type TOCProps,
+} from './slots/toc';
+import { Footer, type FooterProps } from './slots/footer';
+import { Breadcrumb, type BreadcrumbProps } from './slots/breadcrumb';
+import { Container } from './slots/container';
+import type { TOCItemType } from '@hanzo/docs-core/toc';
+import { buttonVariants } from '@/components/ui/button';
+import { Edit } from 'lucide-react';
+>>>>>>> dev
 
-export interface DocsPageProps {
+export interface DocsPageProps extends ComponentProps<'article'> {
   toc?: TOCItemType[];
-  /**
-   * - `multiple` (default): Accept multiple active items
-   * - `single`: Only accept one active item at most
-   * */
-  tocMode?: 'single' | 'multiple';
+
   /**
    * Extend the page to fill all available space
    *
    * @defaultValue false
    */
   full?: boolean;
-  children?: ReactNode;
+  slots?: Partial<DocsPageSlots>;
 
-  TOC?: Renderer<TOCMainProps>;
-  TOCPopover?: Renderer<TOCPopoverProps>;
-  /** Footer navigation, located under the page body. */
-  Footer?: Renderer<FooterProps>;
-  Container?: Renderer<ComponentProps<'div'>>;
-  Breadcrumb?: Renderer<BreadcrumbProps>;
-
-  /** @deprecated use `Footer` instead. */
   footer?: FooterOptions;
-  /** @deprecated use `Container` instead. */
-  className?: string;
-  /** @deprecated use `Breadcrumb` instead. */
   breadcrumb?: BreadcrumbOptions;
-  /** @deprecated use `TOC` instead, or `tocMode` for enabling `single`. */
   tableOfContent?: TableOfContentOptions;
-  /** @deprecated use `TOCPopover` instead. */
   tableOfContentPopover?: TableOfContentPopoverOptions;
 }
 
 interface BreadcrumbOptions extends BreadcrumbProps {
   enabled?: boolean;
+  /**
+   * @deprecated use `slots.breadcrumb` instead.
+   */
   component?: ReactNode;
 }
 
 interface FooterOptions extends FooterProps {
   enabled?: boolean;
+  /**
+   * @deprecated use `slots.footer` instead.
+   */
   component?: ReactNode;
 }
 
-interface TableOfContentOptions extends Pick<AnchorProviderProps, 'single'>, TOCProps {
+interface TableOfContentOptions extends Pick<TOCProviderProps, 'single'>, TOCProps {
   enabled?: boolean;
+  /**
+   * @deprecated use `slots.toc` instead.
+   */
   component?: ReactNode;
 }
 
-interface TableOfContentPopoverOptions extends TOCProps {
+interface TableOfContentPopoverOptions extends TOCPopoverProps {
   enabled?: boolean;
+  /**
+   * @deprecated use `slots.tocPopover` instead.
+   */
   component?: ReactNode;
+}
+
+interface DocsPageSlots {
+  toc: {
+    provider: FC<TOCProviderProps>;
+    main: FC<TOCProps>;
+    popover: FC<TOCPopoverProps>;
+  };
+  container: FC<ComponentProps<'article'>>;
+  footer: FC<FooterProps>;
+  breadcrumb: FC<BreadcrumbProps>;
+}
+
+type PageSlotsProps = Pick<DocsPageProps, 'full'>;
+
+const PageContext = createContext<{
+  props: PageSlotsProps;
+  slots: DocsPageSlots;
+} | null>(null);
+
+export function useDocsPage() {
+  const context = use(PageContext);
+  if (!context)
+    throw new Error(
+      'Please use page components under <DocsPage /> (`@hanzo/docs-ui/layouts/docs/page`).',
+    );
+  return context;
 }
 
 export function DocsPage({
-  tableOfContent: tocProps = {},
-  TOC: TOCRenderer,
-  tableOfContentPopover: tocPopoverProps = {},
-  TOCPopover: TOCPopoverRenderer,
-  footer = {},
-  Footer = footer.enabled === false
-    ? false
-    : footer.component
-      ? new ChildrenRenderer(footer.component)
-      : footer,
-  className,
-  Container = { className },
-  breadcrumb = {},
-  Breadcrumb = breadcrumb.enabled === false
-    ? false
-    : breadcrumb.component
-      ? new ChildrenRenderer(breadcrumb.component)
-      : breadcrumb,
+  tableOfContent: { enabled: tocEnabled, single, ...tocProps } = {},
+  tableOfContentPopover: { enabled: tocPopoverEnabled, ...tocPopoverProps } = {},
+  breadcrumb: { enabled: breadcrumbEnabled = true, ...breadcrumb } = {},
+  footer: { enabled: footerEnabled = true, ...footer } = {},
   full = false,
   toc = [],
-  tocMode = tocProps.single ? 'single' : 'multiple',
+  slots: defaultSlots = {},
   children,
+  ...containerProps
 }: DocsPageProps) {
-  if (!full && (tocProps.enabled ?? (toc.length > 0 || tocProps.footer || tocProps.header))) {
-    TOCRenderer ??= tocProps.component ? new ChildrenRenderer(tocProps.component) : tocProps;
-  } else {
-    // force disable toc in full mode or when no content
-    TOCRenderer = false;
-  }
+  tocEnabled ??= Boolean(!full && (toc.length > 0 || tocProps.footer || tocProps.header));
+  tocPopoverEnabled ??= Boolean(toc.length > 0 || tocPopoverProps.header || tocPopoverProps.footer);
 
-  if (
-    tocPopoverProps.enabled ??
-    (toc.length > 0 || tocPopoverProps.header || tocPopoverProps.footer)
-  ) {
-    TOCPopoverRenderer ??= tocPopoverProps.component
-      ? new ChildrenRenderer(tocPopoverProps.component)
-      : tocPopoverProps;
-  } else {
-    TOCPopoverRenderer = false;
-  }
-
-  const renderBreadcrumb = renderer(Breadcrumb, PageBreadcrumb);
-  const renderFooter = renderer(Footer, PageFooter);
-  const renderContainer = renderer(Container, 'article');
-  const renderToc = renderer(TOCRenderer, TOC);
-  const renderTocPopover = renderer(TOCPopoverRenderer, TOCPopover);
+  const slots: DocsPageSlots = {
+    breadcrumb: defaultSlots.breadcrumb ?? Breadcrumb,
+    footer: defaultSlots.footer ?? Footer,
+    toc: defaultSlots.toc ?? {
+      provider: TOCProvider,
+      main: TOC,
+      popover: TOCPopover,
+    },
+    container: defaultSlots.container ?? Container,
+  };
 
   return (
-    <TOCProvider single={tocMode === 'single'} toc={renderToc || renderTocPopover ? toc : []}>
-      {renderTocPopover?.((t) => t ?? {})}
-      {renderContainer?.((t) => ({
-        id: 'nd-page',
-        'data-full': full,
-        children: (
-          <>
-            {renderBreadcrumb?.((t) => t ?? {})}
-            {children}
-            {renderFooter?.((t) => t ?? {})}
-          </>
-        ),
-        ...t,
-        className: cn(
-          'flex flex-col w-full max-w-[900px] mx-auto [grid-area:main] px-4 py-6 gap-4 md:px-6 md:pt-8 xl:px-8 xl:pt-14',
-          full ? 'max-w-[1168px]' : 'xl:layout:[--fd-toc-width:268px]',
-          t?.className,
-        ),
-      }))}
-      {renderToc?.((t) => t ?? {})}
-    </TOCProvider>
+    <PageContext
+      value={{
+        props: { full },
+        slots,
+      }}
+    >
+      <slots.toc.provider single={single} toc={tocEnabled || tocPopoverEnabled ? toc : []}>
+        {tocPopoverEnabled &&
+          (tocPopoverProps.component ?? <slots.toc.popover {...tocPopoverProps} />)}
+        <slots.container {...containerProps}>
+          {breadcrumbEnabled && (breadcrumb.component ?? <slots.breadcrumb {...breadcrumb} />)}
+          {children}
+          {footerEnabled && (footer.component ?? <slots.footer {...footer} />)}
+        </slots.container>
+        {tocEnabled && (tocProps.component ?? <slots.toc.main {...tocProps} />)}
+      </slots.toc.provider>
+    </PageContext>
   );
 }
 
@@ -155,8 +173,8 @@ export function EditOnGitHub(props: ComponentProps<'a'>) {
         buttonVariants({
           color: 'secondary',
           size: 'sm',
-          className: 'gap-1.5 not-prose',
         }),
+        'gap-1.5 not-prose',
         props.className,
       )}
     >
@@ -200,86 +218,25 @@ export function DocsTitle({ children, className, ...props }: ComponentProps<'h1'
   );
 }
 
-export interface TOCPopoverProps extends TOCProps {
-  container?: ComponentProps<typeof PageTOCPopover>;
-  trigger?: ComponentProps<typeof PageTOCPopoverTrigger>;
-  content?: ComponentProps<typeof PageTOCPopoverContent>;
-}
+export function PageLastUpdate({
+  date: value,
+  ...props
+}: Omit<ComponentProps<'p'>, 'children'> & { date: Date }) {
+  const { text } = useI18n();
+  const [date, setDate] = useState('');
 
-export interface TOCMainProps extends TOCProps {
-  container?: ComponentProps<typeof PageTOCPopover>;
-}
+  useEffect(() => {
+    // to the timezone of client
+    setDate(value.toLocaleDateString());
+  }, [value]);
 
-interface TOCProps {
-  /**
-   * Custom content in TOC container, before the main TOC
-   */
-  header?: ReactNode;
-
-  /**
-   * Custom content in TOC container, after the main TOC
-   */
-  footer?: ReactNode;
-
-  /**
-   * @defaultValue 'normal'
-   */
-  style?: 'normal' | 'clerk';
-}
-
-export function TOC({ container, header, footer, style }: TOCMainProps) {
   return (
-    <div
-      id="nd-toc"
-      {...container}
-      className={cn(
-        'sticky top-(--fd-docs-row-1) h-[calc(var(--fd-docs-height)-var(--fd-docs-row-1))] flex flex-col [grid-area:toc] w-(--fd-toc-width) pt-12 pe-4 pb-2 max-xl:hidden',
-        container?.className,
-      )}
-    >
-      {header}
-      <h3
-        id="toc-title"
-        className="inline-flex items-center gap-1.5 text-sm text-fd-muted-foreground"
-      >
-        <Text className="size-4" />
-        <I18nLabel label="toc" />
-      </h3>
-      <TOCScrollArea>
-        {style === 'clerk' ? <TocClerk.TOCItems /> : <TocDefault.TOCItems />}
-      </TOCScrollArea>
-      {footer}
-    </div>
+    <p {...props} className={cn('text-sm text-fd-muted-foreground', props.className)}>
+      {text.lastUpdate} {date}
+    </p>
   );
 }
 
-export function TOCPopover({
-  container,
-  trigger,
-  content,
-  header,
-  footer,
-  style,
-}: TOCPopoverProps) {
-  return (
-    <PageTOCPopover {...container}>
-      <PageTOCPopoverTrigger {...trigger} />
-      <PageTOCPopoverContent {...content}>
-        {header}
-        <TOCScrollArea>
-          {style === 'clerk' ? <TocClerk.TOCItems /> : <TocDefault.TOCItems />}
-        </TOCScrollArea>
-        {footer}
-      </PageTOCPopoverContent>
-    </PageTOCPopover>
-  );
-}
-
-export {
-  PageLastUpdate,
-  PageBreadcrumb,
-  PageFooter,
-  type BreadcrumbProps,
-  type FooterProps,
-} from './client';
+export { type BreadcrumbProps, Breadcrumb as PageBreadcrumb } from './slots/breadcrumb';
+export { type FooterProps, Footer as PageFooter } from './slots/footer';
 export { MarkdownCopyButton, ViewOptionsPopover } from '@/layouts/shared/page-actions';