+++ packages/cli/src/registry/schema.ts
@@ -8,7 +8,7 @@ export type DownloadedRegistryInfo = z.output<typeof registryInfoSchema>;
 export type File = z.output<typeof fileSchema>;
 export type Component = z.output<typeof componentSchema>;
 
-export const namespaces = ['components', 'lib', 'css', 'route', 'ui', 'block'] as const;
+export const namespaces = ['components', 'lib', 'css', 'route', 'ui', 'layout'] as const;
 
 export const indexSchema = z.object({
   name: z.string(),
@@ -40,26 +40,33 @@ export const componentSchema = z.object({
    * list of sub components, either local (component name) or remote (registry info & component name)
    */
   subComponents: z.array(z.string().or(httpSubComponent)).default([]),
+
+  /**
+   * override variables for the current component & its sub components.
+   *
+   * this is powerful to customise how sub components are installed vs. installing them directly.
+   */
+  variables: z.record(z.string(), z.unknown()).optional(),
 });
 
 export const registryInfoSchema = z.object({
   /**
-   * define used variables, variables can be referenced in the import specifiers of component files.
+   * define metadata for variables, variables can be referenced in the target path of component files, or in plugins.
    */
   variables: z
     .record(
       z.string(),
       z.object({
         description: z.string().optional(),
-        default: z.unknown().optional(),
       }),
     )
     .optional(),
   /**
-   * provide variables to sub components
+   * override variables for all components.
    */
   env: z.record(z.string(), z.unknown()).optional(),
   indexes: z.array(indexSchema).default([]),
+  unlistedIndexes: z.array(indexSchema).default([]),
 
   registries: z.array(z.string()).optional(),
 });