+++ packages/radix-ui/src/components/toc/clerk.tsx
@@ -1,64 +1,78 @@
 'use client';
+<<<<<<< HEAD
 import * as Primitive from '@hanzo/docs-core/toc';
 import { type ComponentProps, useEffect, useRef, useState } from 'react';
+=======
+import * as Primitive from '@hanzo/docs-core/toc';
+import { type ComponentProps, useEffect, useEffectEvent, useRef, useState } from 'react';
+>>>>>>> dev
 import { cn } from '@/utils/cn';
 import { TocThumb, useTOCItems } from '.';
 import { mergeRefs } from '@/utils/merge-refs';
 import { useI18n } from '@/contexts/i18n';
 
+interface ComputedSVG {
+  d: string;
+  width: number;
+  height: number;
+}
+
 export function TOCItems({ ref, className, ...props }: ComponentProps<'div'>) {
   const containerRef = useRef<HTMLDivElement>(null);
   const items = useTOCItems();
   const { text } = useI18n();
+  const [svg, setSvg] = useState<ComputedSVG>();
 
-  const [svg, setSvg] = useState<{
-    path: string;
-    width: number;
-    height: number;
-  }>();
+  const onResize = useEffectEvent(() => {
+    const container = containerRef.current;
+    if (!container || container.clientHeight === 0) return;
+    let w = 0;
+    let h = 0;
+    let b0 = 0;
+    let d = '';
+
+    for (let i = 0; i < items.length; i++) {
+      const element: HTMLElement | null = container.querySelector(
+        `a[href="#${items[i].url.slice(1)}"]`,
+      );
+      if (!element) continue;
+
+      const styles = getComputedStyle(element);
+      const offset = getLineOffset(items[i].depth) + 1,
+        top = element.offsetTop + parseFloat(styles.paddingTop),
+        bottom = element.offsetTop + element.clientHeight - parseFloat(styles.paddingBottom);
+
+      w = Math.max(offset, w);
+      h = Math.max(h, bottom);
+
+      if (i === 0) {
+        d += ` M${offset} ${top} L${offset} ${bottom}`;
+      } else {
+        const pOffset = getLineOffset(items[i - 1].depth) + 1;
+        d += ` C ${pOffset} ${top - 4} ${offset} ${b0! + 4} ${offset} ${top} L${offset} ${bottom}`;
+      }
+
+      b0 = bottom;
+    }
+
+    w += 1;
+    setSvg({
+      d,
+      width: w,
+      height: h,
+    });
+  });
 
   useEffect(() => {
     if (!containerRef.current) return;
-    const container = containerRef.current;
-
-    function onResize(): void {
-      if (container.clientHeight === 0) return;
-      let w = 0,
-        h = 0;
-      const d: string[] = [];
-      for (let i = 0; i < items.length; i++) {
-        const element: HTMLElement | null = container.querySelector(
-          `a[href="#${items[i].url.slice(1)}"]`,
-        );
-        if (!element) continue;
-
-        const styles = getComputedStyle(element);
-        const offset = getLineOffset(items[i].depth) + 1,
-          top = element.offsetTop + parseFloat(styles.paddingTop),
-          bottom = element.offsetTop + element.clientHeight - parseFloat(styles.paddingBottom);
-
-        w = Math.max(offset, w);
-        h = Math.max(h, bottom);
-
-        d.push(`${i === 0 ? 'M' : 'L'}${offset} ${top}`);
-        d.push(`L${offset} ${bottom}`);
-      }
-
-      setSvg({
-        path: d.join(' '),
-        width: w + 1,
-        height: h,
-      });
-    }
-
     const observer = new ResizeObserver(onResize);
     onResize();
 
-    observer.observe(container);
+    observer.observe(containerRef.current);
     return () => {
       observer.disconnect();
     };
-  }, [items]);
+  }, []);
 
   if (items.length === 0)
     return (
@@ -70,24 +84,28 @@ export function TOCItems({ ref, className, ...props }: ComponentProps<'div'>) {
   return (
     <>
       {svg && (
-        <div
-          className="absolute start-0 top-0 rtl:-scale-x-100"
+        <TocThumb
+          containerRef={containerRef}
+          className="absolute top-0 inset-s-0"
           style={{
             width: svg.width,
             height: svg.height,
-            maskImage: `url("data:image/svg+xml,${
-              // Inline SVG
-              encodeURIComponent(
-                `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 ${svg.width} ${svg.height}"><path d="${svg.path}" stroke="black" stroke-width="1" fill="none" /></svg>`,
-              )
-            }")`,
           }}
         >
-          <TocThumb
-            containerRef={containerRef}
-            className="absolute w-full top-(--fd-top) h-(--fd-height) bg-fd-primary transition-[top,height]"
-          />
-        </div>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox={`0 0 ${svg.width} ${svg.height}`}
+            className="absolute transition-[clip-path]"
+            style={{
+              width: svg.width,
+              height: svg.height,
+              clipPath: `polygon(0 var(--fd-top), 100% var(--fd-top), 100% calc(var(--fd-top) + var(--fd-height)), 0 calc(var(--fd-top) + var(--fd-height)))`,
+            }}
+          >
+            <path d={svg.d} className="stroke-fd-primary" strokeWidth="1" fill="none" />
+          </svg>
+          <ThumbBox />
+        </TocThumb>
       )}
       <div ref={mergeRefs(containerRef, ref)} className={cn('flex flex-col', className)} {...props}>
         {items.map((item, i) => (
@@ -103,6 +121,34 @@ export function TOCItems({ ref, className, ...props }: ComponentProps<'div'>) {
   );
 }
 
+function ThumbBox() {
+  const itemInfos = Primitive.useItems();
+  const startIdx = itemInfos.findIndex((info) => info.active);
+  const endIdx = itemInfos.findLastIndex((info) => info.active);
+  if (startIdx === -1) return;
+
+  let lastInactiveIdx = -1;
+  for (let i = 0; i < itemInfos.length; i++) {
+    const item = itemInfos[i];
+    if (item.active) continue;
+    if (lastInactiveIdx === -1 || itemInfos[lastInactiveIdx].t < item.t) {
+      lastInactiveIdx = i;
+    }
+  }
+  const isStart = endIdx < lastInactiveIdx;
+
+  return (
+    <div
+      className="absolute size-1 bg-fd-primary rounded-full transition-transform"
+      style={{
+        translate: `calc(${getLineOffset(itemInfos[isStart ? startIdx : endIdx].original.depth)}px - 1.25px) ${
+          isStart ? 'var(--fd-top)' : 'calc(var(--fd-top) + var(--fd-height))'
+        }`,
+      }}
+    />
+  );
+}
+
 function getItemOffset(depth: number): number {
   if (depth <= 2) return 14;
   if (depth === 3) return 26;
@@ -110,7 +156,9 @@ function getItemOffset(depth: number): number {
 }
 
 function getLineOffset(depth: number): number {
-  return depth >= 3 ? 10 : 0;
+  if (depth <= 2) return 2;
+  if (depth === 3) return 10;
+  return 20;
 }
 
 function TOCItem({
@@ -132,21 +180,25 @@ function TOCItem({
       style={{
         paddingInlineStart: getItemOffset(item.depth),
       }}
-      className="prose relative py-1.5 text-sm text-fd-muted-foreground hover:text-fd-accent-foreground transition-colors wrap-anywhere first:pt-0 last:pb-0 data-[active=true]:text-fd-primary"
+      className="prose relative py-1.5 text-sm scroll-m-4 text-fd-muted-foreground hover:text-fd-accent-foreground transition-colors wrap-anywhere first:pt-0 last:pb-0 data-[active=true]:text-fd-primary"
     >
       {offset !== upperOffset && (
         <svg
           xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 16 16"
-          className="absolute -top-1.5 start-0 size-4 rtl:-scale-x-100"
+          viewBox={`${Math.min(offset, upperOffset)} 0 ${Math.abs(upperOffset - offset)} 12`}
+          className="absolute -top-1.5"
+          style={{
+            width: Math.abs(upperOffset - offset) + 1,
+            height: 12,
+            insetInlineStart: Math.min(offset, upperOffset),
+          }}
         >
-          <line
-            x1={upperOffset}
-            y1="0"
-            x2={offset}
-            y2="12"
-            className="stroke-fd-foreground/10"
+          <path
+            d={`M ${upperOffset} 0 C ${upperOffset} 8 ${offset} 4 ${offset} 12`}
+            stroke="black"
             strokeWidth="1"
+            fill="none"
+            className="stroke-fd-foreground/10"
           />
         </svg>
       )}